+++ CloudServer/src/Core/RequestManager.h
@@ -10,21 +10,21 @@ class RequestManager
 private:
 	//Static member variables
 	//A list of connected clients
-	static std::map<uint32_t, ClientData*> connectedClients;
+	static std::map<uint64_t, ClientData*> connectedClients;
 public:
-	static const std::string NewClientConnected(uint32_t clientSocket);
-	static void ClientDisconnected(uint32_t clientSocket);
+	static const std::string NewClientConnected(uint64_t clientSocket);
+	static void ClientDisconnected(uint64_t clientSocket);
 
-	static const std::string UnknownRequest(uint32_t clientSocket);
-	static const std::string InvalidMessageLength(uint32_t clientSocket, const std::string message);
+	static const std::string UnknownRequest(uint64_t clientSocket);
+	static const std::string InvalidMessageLength(uint64_t clientSocket, const std::string message);
 
-	static const std::string RegisterNewAccount(uint32_t clientSocket, const std::vector<std::string>& messageTokens);
-	static const std::string LoginIntoAccount(uint32_t clientSocket, const std::vector<std::string>& messageTokens);
-	static const std::string ReceivePublicKey(uint32_t clientSocket, const std::string message);
+	static const std::string RegisterNewAccount(uint64_t clientSocket, const std::vector<std::string>& messageTokens);
+	static const std::string LoginIntoAccount(uint64_t clientSocket, const std::vector<std::string>& messageTokens);
+	static const std::string ReceivePublicKey(uint64_t clientSocket, const std::string message);
 
 	//Accessors
-	static bool ClientSupportsEncryption(uint32_t clientSocket);
-	static std::string GetClientSecret(uint32_t clientSocket);
+	static bool ClientSupportsEncryption(uint64_t clientSocket);
+	static std::string GetClientSecret(uint64_t clientSocket);
 
 	//Deallocate the memory for the remaining users if the server closes
 	static void ClearRequestManager();
